drm/radeon: Partial update to Linux 3.12
  This brings new radeon features (mostly code refactoring):
    DMA functions seperated from asics code;
    UVD block separation;
    Improvements to DPMS (by default disabled);
    Better audio handling.

  Notes:
    drm/ttm layer is put into transitional state pending memory functions rewrite.
    Currently VRAM handling capabilities are limited,
      still enough for light gaming and watching videos.
    X acceleration on radeon SI cards requires updated xorg+mesa(glamour support).

+++ sys/dev/drm/include/drm/ttm/ttm_bo_api.h
@@ -37,6 +37,8 @@
 #include <linux/kref.h>
 #include <linux/list.h>
 #include <linux/wait.h>
+#include <linux/mutex.h>
+#include <linux/reservation.h>
 
 struct ttm_bo_device;
 
@@ -233,7 +235,7 @@ struct ttm_buffer_object {
 	struct list_head ddestroy;
 	struct list_head swap;
 	struct list_head io_reserve_lru;
-	uint32_t val_seq;
+	unsigned long val_seq;
 	bool seq_valid;
 
 	/**
